requirements.txt
# Основные зависимости для научных вычислений
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0

# Обработка сигналов и ЭЭГ
mne>=1.0.0
pyedflib>=0.1.0
PyWavelets>=1.1.0

# Машинное обучение
scikit-learn>=1.0.0

# GUI (графический интерфейс)
PyQt5>=5.15.0

# Real-time функциональность
pyserial>=3.5

# Системная информация и производительность
psutil>=5.8.0

# Тестирование
pytest>=6.0.0
pytest-qt>=4.0.0

# Дополнительные зависимости для стабильной работы
setuptools>=45.0.0
wheel>=0.36.0

# Для работы с HDF5 файлами (требуется для MNE)
h5py>=3.1.0

# Для улучшенной работы с matplotlib
pillow>=8.0.0

# Для генерации PDF отчетов
reportlab>=3.6.0
